refactor(studio): extract timeline render contracts
